Feature/유저 소감 말하기 (1차) 작업 (#601)
* User-Contribution UI 개발 (#580)

* style: 메달 아이콘 추가

* feat: 말풍선 컴포넌트 구현

* feat: 덕퀴즈 내 기록 UI WIP

* feat: 내 기록 관련 비즈니스 로직 추가

* feat: 덕키 메달 컴포넌트

* style: UserContribution에 필요한 아이콘 정리

* feat: UI 테스트를 위한 더미 데이터 Mapper에 구현

* feat: 랭킹 UI 스펙에 맞게 수정

* chore: bump quack-v2 2.0.0-alpha04 to 2.0.0-alpha07

* fix: DuckieIcon ImageVector 형식으로 변환

* fix: 바뀐 QuackIcon 형식 대응

* style: detekt 검출사항 적용

* fix: user-contribution QuackIcon 변동사항 대응

* feat: 디테일 화면 스펙에 맞게 업데이트

* feat: 랭킹 구별 extension

* feat: user reaction state

* feat: 유저 결과 화면 스펙에 맞게 변경

* feat: 키보드 UX 개선 & 리컴포지션 개선

* fix: 하드코딩 제거

* Feature/유저 소감 작성하기 비즈니스 로직 (#600)

* fix: BASE_URL이 잘못 들어가는 오류

* fix: change Timber type error to info

* feat: timer 속도 빠르게 변경하고, 서버에서 내려온 값으로 타이머 설정 (#591)

* fix: change time type Int to Double

* fix: 유저 소감 작성하기 UI 디테일 잡기

* feat: update QuizInfoResponse

* style: lint style 적용

* feat: PostQuizReaction 비즈니스 로직 작업

* feat: 이전 기록보다 더 높은 기록일 경우에만 reaction 받기

---------


* feat: 유저 소감 말하기 UI 디테일 잡기

* feat: StartExam 예외처리

* chore: update version 1.1.4 to 1.2.0

* fix: 잘못된 문구 수정

* fix: 도로님 QA 반영

디자인 안 맞는 부분 수정, NotFound 메시지 추가

* feat: 마이페이지 버튼 수정 & 시험 결과 페이지 imePadding 적용

* fix: ExamResultScreen UI 구조 최적화

* fix: change Quiz time type Int to Double

* feat: ReactionTextField Dialog로 처리

* fix: Reaction이 없을 경우 Center처리

* 유저 소감 말하기 qa 4차 (#607)

* feat: 순위 프로필 사진 기준으로 가운데 정렬, 문구 변경

* refactor: ./gradlew build

* feat: reaction 제출 성공 시 toast 처리

* fix: doro님 유저 소감 말하기 5차 QA 반영

* refactor: QuackReactionTextArea 분리

* fix: 덕력고사 생성하기 제거

* chore: bump versionCode 20 to 21

* refactor: MedalColor enum으로 관리

* refactor: 랭킹 extension 활용

* fix: 락스님 리뷰 반영

---------

+++ feature/solve-problem/src/main/kotlin/team/duckie/app/android/feature/solve/problem/viewmodel/sideeffect/SolveProblemSideEffect.kt
@@ -16,7 +16,7 @@ internal sealed class SolveProblemSideEffect {
     class FinishQuiz(
         val examId: Int,
         val correctProblemCount: Int,
-        val time: Int,
+        val time: Double,
         val problemId: Int?,
         val requirementAnswer: String,
         val inputCurrentAnswer: String?,
